Time average

+++ app/src/main/java/com/etsisi/appquitectura/data/model/dto/UserDTO.kt
@@ -9,7 +9,7 @@ data class UserDTO (
         val id: String = String.EMPTY,
         val name: String = String.EMPTY,
         val subject: String = QuestionSubject.UNKNOWN.value,
-        val gameExperience: Int = 0,
+        val gameExperience: Long = 0,
         val totalQuestionsAnswered: Int = 0,
         val totalCorrectQuestionsAnswered: Int = 0
 ): FirestoreDTO() {
